Move unstored_dependencies_graph to query.py; define assemble_dependency_graph
